consumer/requirements.txt
quixstreams==3.14.1
python-dotenv
boto3
pandas
pyarrow
Django>=4.0




